requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Django==4.1.7
djangorestframework==3.14.0
fabric==3.0.0
fake==0.8
Faker==17.0.0
idna==3.4
invoke==2.0.0
paramiko==3.0.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
